// src/app/api/model/apiResponsePanier.ts
/**
 * E-Commerce API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Panier } from './panier';

export interface ApiResponsePanier {
  timestamp: string;
  data: Panier;
  code: number;
  status: string;
  message?: string;
  details?: { [key: string]: any };
}