// src/app/api/model/panier.ts
/**
 * E-Commerce API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PanierProduct } from './panierProduct';

export interface Panier {
  panierCode: string;
  dateCreation?: string;
  clientCode: string;
  produits: Array<PanierProduct>;
  state: string;
}